chore : Move user test_apis.py
